+++ storescp/Cargo.toml
@@ -13,7 +13,7 @@ readme = "README.md"
 [dependencies]
 clap = { version = "4.0.18", features = ["derive"] }
 dicom-core = { path = '../core', version = "0.7.0" }
-dicom-ul = { path = '../ul', version = "0.7.0", features = [] }
+dicom-ul = { path = '../ul', version = "0.7.0" }
 dicom-object = { path = '../object', version = "0.7.0" }
 dicom-encoding = { path = "../encoding/", version = "0.7.0" }
 dicom-dictionary-std = { path = "../dictionary-std/", version = "0.7.0" }
@@ -21,4 +21,8 @@ dicom-transfer-syntax-registry = { path = "../transfer-syntax-registry/", versio
 snafu = "0.8"
 tracing = "0.1.36"
 tracing-subscriber = "0.3.15"
-tokio = "1.38.0"
+tokio = { version = "1.38.0", features = ["full"], optional = true }
+
+[features]
+deafult = ["async"]
+async = ["dicom-ul/async", "dep:tokio"]
