+++ storescp/src/main.rs
@@ -1,22 +1,24 @@
 use std::{
-    net::{Ipv4Addr, SocketAddrV4},
+    net::{Ipv4Addr, SocketAddrV4, TcpListener},
     path::PathBuf,
-    sync::Arc,
 };
 
 use clap::Parser;
 use dicom_core::{dicom_value, DataElement, VR};
 use dicom_dictionary_std::tags;
-use dicom_encoding::transfer_syntax::TransferSyntaxIndex;
-use dicom_object::{FileMetaTableBuilder, InMemDicomObject, StandardDataDictionary};
-use dicom_transfer_syntax_registry::TransferSyntaxRegistry;
-use dicom_ul::{pdu::PDataValueType, Pdu};
-use snafu::{OptionExt, Report, ResultExt, Whatever};
-use tracing::{debug, error, info, warn, Level};
+use dicom_object::{InMemDicomObject, StandardDataDictionary};
+use tracing::{error, info, Level};
 
-use crate::transfer::ABSTRACT_SYNTAXES;
 
 mod transfer;
+#[cfg(feature = "async")]
+mod store_async;
+#[cfg(feature = "async")]
+use store_async::run;
+#[cfg(not(feature = "async"))]
+mod store_sync;
+#[cfg(not(feature = "async"))]
+use store_sync::run;
 
 /// DICOM C-STORE SCP
 #[derive(Debug, Parser)]
@@ -48,267 +50,7 @@ struct App {
     port: u16,
 }
 
-async fn run(scu_stream: tokio::net::TcpStream, args: &App) -> Result<(), Whatever> {
-    let App {
-        verbose,
-        calling_ae_title,
-        strict,
-        uncompressed_only,
-        promiscuous,
-        max_pdu_length,
-        out_dir,
-        port: _,
-    } = args;
-    let verbose = *verbose;
 
-    let mut buffer: Vec<u8> = Vec::with_capacity(*max_pdu_length as usize);
-    let mut instance_buffer: Vec<u8> = Vec::with_capacity(1024 * 1024);
-    let mut msgid = 1;
-    let mut sop_class_uid = "".to_string();
-    let mut sop_instance_uid = "".to_string();
-
-    let mut options = dicom_ul::association::ServerAssociationOptions::new()
-        .accept_any()
-        .ae_title(calling_ae_title)
-        .strict(*strict)
-        .promiscuous(*promiscuous);
-
-    if *uncompressed_only {
-        options = options
-            .with_transfer_syntax("1.2.840.10008.1.2")
-            .with_transfer_syntax("1.2.840.10008.1.2.1");
-    } else {
-        for ts in TransferSyntaxRegistry.iter() {
-            if !ts.is_unsupported() {
-                options = options.with_transfer_syntax(ts.uid());
-            }
-        }
-    };
-
-    for uid in ABSTRACT_SYNTAXES {
-        options = options.with_abstract_syntax(*uid);
-    }
-
-    let mut association = options
-        .establish(scu_stream)
-        .await
-        .whatever_context("could not establish association")?;
-
-    info!("New association from {}", association.client_ae_title());
-    debug!(
-        "> Presentation contexts: {:?}",
-        association.presentation_contexts()
-    );
-
-    loop {
-        match association.receive().await {
-            Ok(mut pdu) => {
-                if verbose {
-                    debug!("scu ----> scp: {}", pdu.short_description());
-                }
-                match pdu {
-                    Pdu::PData { ref mut data } => {
-                        if data.is_empty() {
-                            debug!("Ignoring empty PData PDU");
-                            continue;
-                        }
-
-                        for data_value in data {
-                            if data_value.value_type == PDataValueType::Data && !data_value.is_last
-                            {
-                                instance_buffer.append(&mut data_value.data);
-                            } else if data_value.value_type == PDataValueType::Command
-                                && data_value.is_last
-                            {
-                                // commands are always in implict VR LE
-                                let ts =
-                                    dicom_transfer_syntax_registry::entries::IMPLICIT_VR_LITTLE_ENDIAN
-                                        .erased();
-                                let data_value = &data_value;
-                                let v = &data_value.data;
-
-                                let obj = InMemDicomObject::read_dataset_with_ts(v.as_slice(), &ts)
-                                    .whatever_context("failed to read incoming DICOM command")?;
-                                let command_field = obj
-                                    .element(tags::COMMAND_FIELD)
-                                    .whatever_context("Missing Command Field")?
-                                    .uint16()
-                                    .whatever_context("Command Field is not an integer")?;
-
-                                if command_field == 0x0030 {
-                                    // Handle C-ECHO-RQ
-                                    let cecho_response = create_cecho_response(msgid);
-                                    let mut cecho_data = Vec::new();
-
-                                    cecho_response
-                                        .write_dataset_with_ts(&mut cecho_data, &ts)
-                                        .whatever_context(
-                                            "could not write C-ECHO response object",
-                                        )?;
-
-                                    let pdu_response = Pdu::PData {
-                                        data: vec![dicom_ul::pdu::PDataValue {
-                                            presentation_context_id: data_value
-                                                .presentation_context_id,
-                                            value_type: PDataValueType::Command,
-                                            is_last: true,
-                                            data: cecho_data,
-                                        }],
-                                    };
-                                    association.send(&pdu_response).await.whatever_context(
-                                        "failed to send C-ECHO response object to SCU",
-                                    )?;
-                                } else {
-                                    msgid = obj
-                                        .element(tags::MESSAGE_ID)
-                                        .whatever_context("Missing Message ID")?
-                                        .to_int()
-                                        .whatever_context("Message ID is not an integer")?;
-                                    sop_class_uid = obj
-                                        .element(tags::AFFECTED_SOP_CLASS_UID)
-                                        .whatever_context("missing Affected SOP Class UID")?
-                                        .to_str()
-                                        .whatever_context(
-                                            "could not retrieve Affected SOP Class UID",
-                                        )?
-                                        .to_string();
-                                    sop_instance_uid = obj
-                                        .element(tags::AFFECTED_SOP_INSTANCE_UID)
-                                        .whatever_context("missing Affected SOP Instance UID")?
-                                        .to_str()
-                                        .whatever_context(
-                                            "could not retrieve Affected SOP Instance UID",
-                                        )?
-                                        .to_string();
-                                }
-                                instance_buffer.clear();
-                            } else if data_value.value_type == PDataValueType::Data
-                                && data_value.is_last
-                            {
-                                instance_buffer.append(&mut data_value.data);
-
-                                let presentation_context = association
-                                    .presentation_contexts()
-                                    .iter()
-                                    .find(|pc| pc.id == data_value.presentation_context_id)
-                                    .whatever_context("missing presentation context")?;
-                                let ts = &presentation_context.transfer_syntax;
-
-                                let obj = InMemDicomObject::read_dataset_with_ts(
-                                    instance_buffer.as_slice(),
-                                    TransferSyntaxRegistry.get(ts).unwrap(),
-                                )
-                                .whatever_context("failed to read DICOM data object")?;
-                                let file_meta = FileMetaTableBuilder::new()
-                                    .media_storage_sop_class_uid(
-                                        obj.element(tags::SOP_CLASS_UID)
-                                            .whatever_context("missing SOP Class UID")?
-                                            .to_str()
-                                            .whatever_context("could not retrieve SOP Class UID")?,
-                                    )
-                                    .media_storage_sop_instance_uid(
-                                        obj.element(tags::SOP_INSTANCE_UID)
-                                            .whatever_context("missing SOP Instance UID")?
-                                            .to_str()
-                                            .whatever_context("missing SOP Instance UID")?,
-                                    )
-                                    .transfer_syntax(ts)
-                                    .build()
-                                    .whatever_context(
-                                        "failed to build DICOM meta file information",
-                                    )?;
-                                let file_obj = obj.with_exact_meta(file_meta);
-
-                                // write the files to the current directory with their SOPInstanceUID as filenames
-                                let mut file_path = out_dir.clone();
-                                file_path.push(
-                                    sop_instance_uid.trim_end_matches('\0').to_string() + ".dcm",
-                                );
-                                file_obj
-                                    .write_to_file(&file_path)
-                                    .whatever_context("could not save DICOM object to file")?;
-                                info!("Stored {}", file_path.display());
-
-                                // send C-STORE-RSP object
-                                // commands are always in implict VR LE
-                                let ts =
-                                    dicom_transfer_syntax_registry::entries::IMPLICIT_VR_LITTLE_ENDIAN
-                                        .erased();
-
-                                let obj = create_cstore_response(
-                                    msgid,
-                                    &sop_class_uid,
-                                    &sop_instance_uid,
-                                );
-
-                                let mut obj_data = Vec::new();
-
-                                obj.write_dataset_with_ts(&mut obj_data, &ts)
-                                    .whatever_context("could not write response object")?;
-
-                                let pdu_response = Pdu::PData {
-                                    data: vec![dicom_ul::pdu::PDataValue {
-                                        presentation_context_id: data_value.presentation_context_id,
-                                        value_type: PDataValueType::Command,
-                                        is_last: true,
-                                        data: obj_data,
-                                    }],
-                                };
-                                association
-                                    .send(&pdu_response)
-                                    .await
-                                    .whatever_context("failed to send response object to SCU")?;
-                            }
-                        }
-                    }
-                    Pdu::ReleaseRQ => {
-                        buffer.clear();
-                        association.send(&Pdu::ReleaseRP).await.unwrap_or_else(|e| {
-                            warn!(
-                                "Failed to send association release message to SCU: {}",
-                                snafu::Report::from_error(e)
-                            );
-                        });
-                        info!(
-                            "Released association with {}",
-                            association.client_ae_title()
-                        );
-                        break;
-                    }
-                    Pdu::AbortRQ { source } => {
-                        warn!("Aborted connection from: {:?}", source);
-                        break;
-                    }
-                    _ => {}
-                }
-            }
-            Err(err @ dicom_ul::association::server::Error::Receive { .. }) => {
-                if verbose {
-                    info!("{}", Report::from_error(err));
-                } else {
-                    info!("{}", err);
-                }
-                break;
-            }
-            Err(err) => {
-                warn!("Unexpected error: {}", Report::from_error(err));
-                break;
-            }
-        }
-    }
-
-    if let Ok(peer_addr) = association.inner_stream().peer_addr() {
-        info!(
-            "Dropping connection with {} ({})",
-            association.client_ae_title(),
-            peer_addr
-        );
-    } else {
-        info!("Dropping connection with {}", association.client_ae_title());
-    }
-
-    Ok(())
-}
 
 fn create_cstore_response(
     message_id: u16,
@@ -358,8 +100,10 @@ fn create_cecho_response(message_id: u16) -> InMemDicomObject<StandardDataDictio
     ])
 }
 
+#[cfg(feature = "async")]
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
+    use std::sync::Arc;
     let args = Arc::new(App::parse());
 
     tracing::subscriber::set_global_default(
@@ -398,6 +142,56 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                 error!("{}", snafu::Report::from_error(e));
             }
         });
+    }
+
+    Ok(())
+}
+
+#[cfg(not(feature = "async"))]
+fn main() -> Result<(), Box<dyn std::error::Error>> {
+    use std::io::Read;
+
+    let args = App::parse();
+
+    tracing::subscriber::set_global_default(
+        tracing_subscriber::FmtSubscriber::builder()
+            .with_max_level(if args.verbose {
+                Level::DEBUG
+            } else {
+                Level::INFO
+            })
+            .finish(),
+    )
+    .unwrap_or_else(|e| {
+        eprintln!(
+            "Could not set up global logger: {}",
+            snafu::Report::from_error(e)
+        );
+    });
+
+    std::fs::create_dir_all(&args.out_dir).unwrap_or_else(|e| {
+        error!("Could not create output directory: {}", e);
+        std::process::exit(-2);
+    });
+
+    let listen_addr = SocketAddrV4::new(Ipv4Addr::from(0), args.port);
+    let listener = TcpListener::bind(listen_addr)?;
+    info!(
+        "{} listening on: tcp://{}",
+        &args.calling_ae_title, listen_addr
+    );
+
+    for stream in listener.incoming() {
+        match stream {
+            Ok(scu_stream) => {
+                if let Err(e) = run(scu_stream, &args) {
+                    error!("{}", snafu::Report::from_error(e));
+                }
+            }
+            Err(e) => {
+                error!("{}", snafu::Report::from_error(e));
+            }
+        }
     }
 
     Ok(())
