+++ storescu/Cargo.toml
@@ -14,6 +14,7 @@ readme = "README.md"
 default = ["transcode"]
 # support DICOM transcoding
 transcode = ["dep:dicom-pixeldata"]
+async = ["dicom-ul/async", "dep:tokio"]
 
 [dependencies]
 clap = { version  = "4.0.18", features = ["derive"] }
@@ -29,3 +30,4 @@ indicatif = "0.17.0"
 tracing = "0.1.34"
 tracing-subscriber = "0.3.11"
 snafu = "0.8"
+tokio = { version = "1.38.0", features = ["full"], optional = true }
