+++ storescu/src/main.rs
@@ -3,23 +3,23 @@ use dicom_core::{dicom_value, header::Tag, DataElement, VR};
 use dicom_dictionary_std::{tags, uids};
 use dicom_encoding::transfer_syntax;
 use dicom_encoding::TransferSyntax;
-use dicom_object::{mem::InMemDicomObject, open_file, DefaultDicomObject, StandardDataDictionary};
+use dicom_object::{mem::InMemDicomObject, DefaultDicomObject, StandardDataDictionary};
 use dicom_transfer_syntax_registry::TransferSyntaxRegistry;
-use dicom_ul::{
-    association::ClientAssociationOptions,
-    pdu::{PDataValue, PDataValueType, Pdu},
-};
 use indicatif::{ProgressBar, ProgressStyle};
 use snafu::prelude::*;
 use snafu::{Report, Whatever};
 use std::collections::HashSet;
 use std::ffi::OsStr;
-use std::io::Write;
 use std::path::{Path, PathBuf};
 use std::time::Duration;
 use tracing::{debug, error, info, warn, Level};
 use transfer_syntax::TransferSyntaxIndex;
 use walkdir::WalkDir;
+
+#[cfg(not(feature = "async"))]
+mod store_sync;
+#[cfg(feature = "async")]
+mod store_async;
 
 /// DICOM C-STORE SCU
 #[derive(Debug, Parser)]
@@ -131,6 +131,7 @@ enum Error {
     },
 }
 
+#[cfg(not(feature = "async"))]
 fn main() {
     run().unwrap_or_else(|e| {
         error!("{}", Report::from_error(e));
@@ -138,39 +139,18 @@ fn main() {
     });
 }
 
-fn run() -> Result<(), Error> {
-    let App {
-        addr,
-        files,
-        verbose,
-        message_id,
-        calling_ae_title,
-        called_ae_title,
-        max_pdu_length,
-        fail_first,
-        mut never_transcode,
-        username,
-        password,
-        kerberos_service_ticket,
-        saml_assertion,
-        jwt,
-    } = App::parse();
-
-    // never transcode if the feature is disabled
-    if cfg!(not(feature = "transcode")) {
-        never_transcode = true;
-    }
-
-    tracing::subscriber::set_global_default(
-        tracing_subscriber::FmtSubscriber::builder()
-            .with_max_level(if verbose { Level::DEBUG } else { Level::INFO })
-            .finish(),
-    )
-    .whatever_context("Could not set up global logging subscriber")
-    .unwrap_or_else(|e: Whatever| {
-        eprintln!("[ERROR] {}", Report::from_error(e));
+#[cfg(feature = "async")]
+#[tokio::main]
+async fn main() {
+    run().await.unwrap_or_else(|e| {
+        error!("{}", Report::from_error(e));
+        std::process::exit(-2);
     });
+}
 
+fn check_files(
+    files: Vec<PathBuf>, verbose: bool, never_transcode: bool
+) -> (Vec<DicomFile>, HashSet<(String, String)>){
     let mut checked_files: Vec<PathBuf> = vec![];
     let mut dicom_files: Vec<DicomFile> = vec![];
     let mut presentation_contexts = HashSet::new();
@@ -227,44 +207,63 @@ fn run() -> Result<(), Error> {
         eprintln!("No supported files to transfer");
         std::process::exit(-1);
     }
+    return (dicom_files, presentation_contexts)
+
+
+}
+
+#[cfg(not(feature = "async"))]
+fn run() -> Result<(), Error> {
+    let App {
+        addr,
+        files,
+        verbose,
+        message_id,
+        calling_ae_title,
+        called_ae_title,
+        max_pdu_length,
+        fail_first,
+        mut never_transcode,
+        username,
+        password,
+        kerberos_service_ticket,
+        saml_assertion,
+        jwt,
+    } = App::parse();
+
+    // never transcode if the feature is disabled
+    if cfg!(not(feature = "transcode")) {
+        never_transcode = true;
+    }
+
+    tracing::subscriber::set_global_default(
+        tracing_subscriber::FmtSubscriber::builder()
+            .with_max_level(if verbose { Level::DEBUG } else { Level::INFO })
+            .finish(),
+    )
+    .whatever_context("Could not set up global logging subscriber")
+    .unwrap_or_else(|e: Whatever| {
+        eprintln!("[ERROR] {}", Report::from_error(e));
+    });
 
     if verbose {
         info!("Establishing association with '{}'...", &addr);
     }
+    let (mut dicom_files, presentation_contexts) = check_files(files, verbose, never_transcode);
 
-    let mut scu_init = ClientAssociationOptions::new()
-        .calling_ae_title(calling_ae_title)
-        .max_pdu_length(max_pdu_length);
 
-    for (storage_sop_class_uid, transfer_syntax) in &presentation_contexts {
-        scu_init = scu_init.with_presentation_context(storage_sop_class_uid, vec![transfer_syntax]);
-    }
-
-    if let Some(called_ae_title) = called_ae_title {
-        scu_init = scu_init.called_ae_title(called_ae_title);
-    }
-
-    if let Some(username) = username {
-        scu_init = scu_init.username(username);
-    }
-
-    if let Some(password) = password {
-        scu_init = scu_init.password(password);
-    }
-
-    if let Some(kerberos_service_ticket) = kerberos_service_ticket {
-        scu_init = scu_init.kerberos_service_ticket(kerberos_service_ticket);
-    }
-
-    if let Some(saml_assertion) = saml_assertion {
-        scu_init = scu_init.saml_assertion(saml_assertion);
-    }
-
-    if let Some(jwt) = jwt {
-        scu_init = scu_init.jwt(jwt);
-    }
-
-    let mut scu = scu_init.establish_with(&addr).context(InitScuSnafu)?;
+    let mut scu = get_scu(
+        addr,
+        calling_ae_title,
+        called_ae_title,
+        max_pdu_length,
+        username,
+        password,
+        kerberos_service_ticket,
+        saml_assertion,
+        jwt,
+        presentation_contexts
+    )?;
 
     if verbose {
         info!("Association established");
@@ -313,179 +312,8 @@ fn run() -> Result<(), Error> {
     }
 
     for file in dicom_files {
-        if let (Some(pc_selected), Some(ts_uid_selected)) = (file.pc_selected, file.ts_selected) {
-            if let Some(pb) = &progress_bar {
-                pb.set_message(file.sop_instance_uid.clone());
-            }
-            let cmd = store_req_command(&file.sop_class_uid, &file.sop_instance_uid, message_id);
-
-            let mut cmd_data = Vec::with_capacity(128);
-            cmd.write_dataset_with_ts(
-                &mut cmd_data,
-                &dicom_transfer_syntax_registry::entries::IMPLICIT_VR_LITTLE_ENDIAN.erased(),
-            )
-            .map_err(Box::from)
-            .context(CreateCommandSnafu)?;
-
-            let mut object_data = Vec::with_capacity(2048);
-            let dicom_file =
-                open_file(&file.file).whatever_context("Could not open listed DICOM file")?;
-            let ts_selected = TransferSyntaxRegistry
-                .get(&ts_uid_selected)
-                .with_context(|| UnsupportedFileTransferSyntaxSnafu {
-                    uid: ts_uid_selected.to_string(),
-                })?;
-
-            // transcode file if necessary
-            let dicom_file = into_ts(dicom_file, ts_selected, verbose)?;
-
-            dicom_file
-                .write_dataset_with_ts(&mut object_data, ts_selected)
-                .whatever_context("Could not write object dataset")?;
-
-            let nbytes = cmd_data.len() + object_data.len();
-
-            if verbose {
-                info!(
-                    "Sending file {} (~ {} kB), uid={}, sop={}, ts={}",
-                    file.file.display(),
-                    nbytes / 1_000,
-                    &file.sop_instance_uid,
-                    &file.sop_class_uid,
-                    ts_uid_selected,
-                );
-            }
-
-            if nbytes < scu.acceptor_max_pdu_length().saturating_sub(100) as usize {
-                let pdu = Pdu::PData {
-                    data: vec![
-                        PDataValue {
-                            presentation_context_id: pc_selected.id,
-                            value_type: PDataValueType::Command,
-                            is_last: true,
-                            data: cmd_data,
-                        },
-                        PDataValue {
-                            presentation_context_id: pc_selected.id,
-                            value_type: PDataValueType::Data,
-                            is_last: true,
-                            data: object_data,
-                        },
-                    ],
-                };
-
-                scu.send(&pdu)
-                    .whatever_context("Failed to send C-STORE-RQ")?;
-            } else {
-                let pdu = Pdu::PData {
-                    data: vec![PDataValue {
-                        presentation_context_id: pc_selected.id,
-                        value_type: PDataValueType::Command,
-                        is_last: true,
-                        data: cmd_data,
-                    }],
-                };
-
-                scu.send(&pdu)
-                    .whatever_context("Failed to send C-STORE-RQ command")?;
-
-                {
-                    let mut pdata = scu.send_pdata(pc_selected.id);
-                    pdata
-                        .write_all(&object_data)
-                        .whatever_context("Failed to send C-STORE-RQ P-Data")?;
-                }
-            }
-
-            if verbose {
-                debug!("Awaiting response...");
-            }
-
-            let rsp_pdu = scu
-                .receive()
-                .whatever_context("Failed to receive C-STORE-RSP")?;
-
-            match rsp_pdu {
-                Pdu::PData { data } => {
-                    let data_value = &data[0];
-
-                    let cmd_obj = InMemDicomObject::read_dataset_with_ts(
-                        &data_value.data[..],
-                        &dicom_transfer_syntax_registry::entries::IMPLICIT_VR_LITTLE_ENDIAN
-                            .erased(),
-                    )
-                    .whatever_context("Could not read response from SCP")?;
-                    if verbose {
-                        debug!("Full response: {:?}", cmd_obj);
-                    }
-                    let status = cmd_obj
-                        .element(tags::STATUS)
-                        .whatever_context("Could not find status code in response")?
-                        .to_int::<u16>()
-                        .whatever_context("Status code in response is not a valid integer")?;
-                    let storage_sop_instance_uid = file
-                        .sop_instance_uid
-                        .trim_end_matches(|c: char| c.is_whitespace() || c == '\0');
-
-                    match status {
-                        // Success
-                        0 => {
-                            if verbose {
-                                info!("Successfully stored instance {}", storage_sop_instance_uid);
-                            }
-                        }
-                        // Warning
-                        1 | 0x0107 | 0x0116 | 0xB000..=0xBFFF => {
-                            warn!(
-                                "Possible issue storing instance `{}` (status code {:04X}H)",
-                                storage_sop_instance_uid, status
-                            );
-                        }
-                        0xFF00 | 0xFF01 => {
-                            warn!(
-                                "Possible issue storing instance `{}`: status is pending (status code {:04X}H)",
-                                storage_sop_instance_uid, status
-                            );
-                        }
-                        0xFE00 => {
-                            error!(
-                                "Could not store instance `{}`: operation cancelled",
-                                storage_sop_instance_uid
-                            );
-                            if fail_first {
-                                let _ = scu.abort();
-                                std::process::exit(-2);
-                            }
-                        }
-                        _ => {
-                            error!(
-                                "Failed to store instance `{}` (status code {:04X}H)",
-                                storage_sop_instance_uid, status
-                            );
-                            if fail_first {
-                                let _ = scu.abort();
-                                std::process::exit(-2);
-                            }
-                        }
-                    }
-                }
-
-                pdu @ Pdu::Unknown { .. }
-                | pdu @ Pdu::AssociationRQ { .. }
-                | pdu @ Pdu::AssociationAC { .. }
-                | pdu @ Pdu::AssociationRJ { .. }
-                | pdu @ Pdu::ReleaseRQ
-                | pdu @ Pdu::ReleaseRP
-                | pdu @ Pdu::AbortRQ { .. } => {
-                    error!("Unexpected SCP response: {:?}", pdu);
-                    let _ = scu.abort();
-                    std::process::exit(-2);
-                }
-            }
-        }
-        if let Some(pb) = progress_bar.as_ref() {
-            pb.inc(1)
-        };
+        // TODO
+        scu = store_sync::send_file(scu, file, message_id, progress_bar.as_ref(), verbose, fail_first)?;
     }
 
     if let Some(pb) = progress_bar {
@@ -497,6 +325,121 @@ fn run() -> Result<(), Error> {
     Ok(())
 }
 
+#[cfg(feature = "async")]
+async fn run() -> Result<(), Error> {
+    use crate::store_async::{get_scu, send_file};
+    let App {
+        addr,
+        files,
+        verbose,
+        message_id,
+        calling_ae_title,
+        called_ae_title,
+        max_pdu_length,
+        fail_first,
+        mut never_transcode,
+        username,
+        password,
+        kerberos_service_ticket,
+        saml_assertion,
+        jwt,
+    } = App::parse();
+
+    // never transcode if the feature is disabled
+    if cfg!(not(feature = "transcode")) {
+        never_transcode = true;
+    }
+
+    tracing::subscriber::set_global_default(
+        tracing_subscriber::FmtSubscriber::builder()
+            .with_max_level(if verbose { Level::DEBUG } else { Level::INFO })
+            .finish(),
+    )
+    .whatever_context("Could not set up global logging subscriber")
+    .unwrap_or_else(|e: Whatever| {
+        eprintln!("[ERROR] {}", Report::from_error(e));
+    });
+
+    if verbose {
+        info!("Establishing association with '{}'...", &addr);
+    }
+    let (mut dicom_files, presentation_contexts) = tokio::task::spawn_blocking(
+        move || check_files(files, verbose, never_transcode)
+    ).await.unwrap();
+
+
+    let mut scu = get_scu(
+        addr,
+        calling_ae_title,
+        called_ae_title,
+        max_pdu_length,
+        username,
+        password,
+        kerberos_service_ticket,
+        saml_assertion,
+        jwt,
+        presentation_contexts
+    ).await?;
+
+    if verbose {
+        info!("Association established");
+    }
+
+    for file in &mut dicom_files {
+        // identify the right transfer syntax to use
+        let r: Result<_, Error> =
+            check_presentation_contexts(file, scu.presentation_contexts(), never_transcode)
+                .whatever_context::<_, _>("Could not choose a transfer syntax");
+        match r {
+            Ok((pc, ts)) => {
+                if verbose {
+                    debug!(
+                        "{}: Selected presentation context: {:?}",
+                        file.file.display(),
+                        pc
+                    );
+                }
+                file.pc_selected = Some(pc);
+                file.ts_selected = Some(ts);
+            }
+            Err(e) => {
+                error!("{}", Report::from_error(e));
+                if fail_first {
+                    let _ = scu.abort();
+                    std::process::exit(-2);
+                }
+            }
+        }
+    }
+
+    let progress_bar;
+    if !verbose {
+        progress_bar = Some(ProgressBar::new(dicom_files.len() as u64));
+        if let Some(pb) = progress_bar.as_ref() {
+            pb.set_style(
+                ProgressStyle::default_bar()
+                    .template("[{elapsed_precise}] {bar:40} {pos}/{len} {wide_msg}")
+                    .expect("Invalid progress bar template"),
+            );
+            pb.enable_steady_tick(Duration::new(0, 480_000_000));
+        };
+    } else {
+        progress_bar = None;
+    }
+
+    for file in dicom_files {
+        // TODO
+        scu = send_file(scu, file, message_id, progress_bar.as_ref(), verbose, fail_first).await?;
+    }
+
+    if let Some(pb) = progress_bar {
+        pb.finish_with_message("done")
+    };
+
+    scu.release().await
+        .whatever_context("Failed to release SCU association")?;
+    Ok(())
+}
 fn store_req_command(
     storage_sop_class_uid: &str,
     storage_sop_instance_uid: &str,
