+++ ul/src/association/mod.rs
@@ -22,6 +22,9 @@ mod uid;
 pub(crate) mod pdata;
 
 pub use client::{ClientAssociation, ClientAssociationOptions};
+#[cfg(not(feature = "async"))]
 pub use pdata::{PDataReader, PDataWriter};
+#[cfg(feature = "async")]
+pub use pdata::{AsyncPDataWriter as PDataWriter, AsyncPDataReader as PDataReader};
 pub use server::{ServerAssociation, ServerAssociationOptions};
